+++ packages/devnext/src/pages/_app.tsx
@@ -34,6 +34,7 @@ const WithSDKConfig = ({ children }: { children: React.ReactNode }) => {
         readonlyRPCMap: {
           '0x539': process.env.NEXT_PUBLIC_PROVIDER_RPCURL ?? '',
         },
+        extensionOnly: false,
         logging: {
           developerMode: true,
           sdk: true,