+++ packages/sdk-communication-layer/src/types/TrackingEvent.ts
@@ -10,6 +10,8 @@ export enum TrackingEvents {
   DISCONNECTED = 'sdk_disconnected',
   SDK_USE_EXTENSION = 'sdk_use_extension',
   SDK_RPC_REQUEST = 'sdk_rpc_request',
+  SDK_RPC_REQUEST_RECEIVED = 'sdk_rpc_request_received',
+  SDK_RPC_REQUEST_DONE = 'sdk_rpc_request_done',
   SDK_EXTENSION_UTILIZED = 'sdk_extension_utilized',
   SDK_USE_INAPP_BROWSER = 'sdk_use_inapp_browser',
 }