+++ packages/sdk/src/sdk.ts
@@ -1,8 +1,11 @@
+import { MetaMaskInpageProvider } from '@metamask/providers';
 import {
   CommunicationLayerPreference,
   DappMetadata,
   StorageManagerProps,
 } from '@metamask/sdk-communication-layer';
+import debug from 'debug';
+
 import EventEmitter2 from 'eventemitter2';
 import { createInstance, i18n } from 'i18next';
 import packageJson from '../package.json';
@@ -16,6 +19,7 @@ import {
   terminate,
 } from './services/MetaMaskSDK/ConnectionManager';
 import { connectAndSign } from './services/MetaMaskSDK/ConnectionManager/connectAndSign';
+import { connectWith } from './services/MetaMaskSDK/ConnectionManager/connectWith';
 import { initializeMetaMaskSDK } from './services/MetaMaskSDK/InitializerManager';
 import { RPC_URLS_MAP } from './services/MetaMaskSDK/InitializerManager/setupReadOnlyRPCProviders';
 import {
@@ -25,7 +29,6 @@ import {
 import { SDKLoggingOptions } from './types/SDKLoggingOptions';
 import { SDKUIOptions } from './types/SDKUIOptions';
 import { WakeLockStatus } from './types/WakeLockStatus';
-import { connectWith } from './services/MetaMaskSDK/ConnectionManager/connectWith';
 import { logger } from './utils/logger';
 
 export interface MetaMaskSDKOptions {
@@ -72,6 +75,7 @@ export interface MetaMaskSDKOptions {
 
   /**
    * If true, the SDK will prefer the desktop version of MetaMask over the mobile version.
+   * @deprecated use `extensionOnly` instead
    */
   preferDesktop?: boolean;
 
@@ -181,9 +185,11 @@ export class MetaMaskSDK extends EventEmitter2 {
 
   public extensionActive = false;
 
+  public extension: MetaMaskInpageProvider | undefined;
+
   public _initialized = false;
 
-  public sdkInitPromise?: Promise<void>;
+  public sdkInitPromise?: Promise<void> | undefined = undefined;
 
   public debug = false;
 
@@ -217,7 +223,15 @@ export class MetaMaskSDK extends EventEmitter2 {
     },
   ) {
     super();
+    debug.disable(); // initially disabled
 
+    const developerMode = options.logging?.developerMode === true;
+    const debugEnabled = options.logging?.sdk || developerMode;
+
+    if (debugEnabled) {
+      debug.enable('MM_SDK');
+    }
+    logger(`[MetaMaskSDK: constructor()]: begin.`);
     this.setMaxListeners(50);
 
     if (!options.dappMetadata?.name && !options.dappMetadata?.url) {
@@ -248,12 +262,19 @@ export class MetaMaskSDK extends EventEmitter2 {
       this.defaultReadOnlyChainId = options.defaultReadOnlyChainId.toString();
     }
 
+    // Remove deprecated options
+    if (options.preferDesktop) {
+      console.warn(
+        `The 'preferDesktop' option is deprecated. Use 'extensionOnly' instead.`,
+      );
+    }
+
     this.options = options;
 
     // Automatically initialize the SDK to keep the same behavior as before
     this.init()
       .then(() => {
-        logger(`[MetaMaskSDK: constructor()]: initialized`);
+        logger(`[MetaMaskSDK: constructor()]: initialized successfully.`);
         if (typeof window !== 'undefined') {
           window.mmsdk = this;
         }